Clarify that cores process all key-value pairs, whether or not supported
The core tests use many key-value pairs with keys that are not defined
as "Supported Pairs" in the specification.  Clarify that cores must
report all key-value pairs, regardless of whether those pairs are
listed in the spec as supported.  It is up to the plugin to ignore keys
it does not recognize.