Clarify that cores report all keys, whether or not supported (#37)
* Terminology: mention key-value pairs

- Clarify that cores produce, and plugins consume, key-value pairs.
- Clarify that a plugin should ignore invalid/unsupported values,
  regardless of whether the plugin recognizes that key.
- Bump version to 0.15.1.  This is a PATCH level change since it
  documents existing behaviour more clearly.

* Clarify that cores process all key-value pairs, whether or not supported

The core tests use many key-value pairs with keys that are not defined
as "Supported Pairs" in the specification.  Clarify that cores must
report all key-value pairs, regardless of whether those pairs are
listed in the spec as supported.  It is up to the plugin to ignore keys
it does not recognize.

* Rename "Terms" to "Parts of an EditorConfig file"

To distinguish that section from "Terminology"