Clarify the case sensitivity of .editorconfig file
This has been generally assumed since we don't specifically mention other
capitalizations, and core libraries have also assumed so. However, some
communities seem to have used capital letters in some places, and it's best if
we can clarify.

If we go the other route, as a standard, the current core libraries would break
because they don't work consistently with each other. Additionally, core
libraries are cumbersome to implement on case-sensitive filesystems, because
they always have to use some regex to search the `.editorconfig` file in a
directory, instead of just pointing to the file named `.editorconfig`.

Close editorconfig/editorconfig#523